requirements-dev.txt
pytest==4.5.0
Django==2.2
djangorestframework==3.8.0
